Completed project, all known bugs fixed.

+++ SodaMachine/SodaMachine/Coin.cs
@@ -8,6 +8,6 @@ namespace SodaMachine
 {
     class Coin
     {
-        public double coinValue;
+        public decimal coinValue;
     }
 }
